+++ tests/transformers/ernie4_5/test_modeling.py
@@ -14,18 +14,19 @@
 # limitations under the License.
 from __future__ import annotations
 
-# import tempfile
+import tempfile
 import unittest
 
-# import numpy as np
+import numpy as np
 import paddle
+from parameterized import parameterized
 
 from paddleformers.transformers import (
     Ernie4_5Config,
     Ernie4_5ForCausalLM,
     Ernie4_5Model,
 )
-from tests.testing_utils import slow  # require_package,
+from tests.testing_utils import require_package, slow
 from tests.transformers.test_configuration_common import ConfigTester
 from tests.transformers.test_generation_utils import GenerationTesterMixin
 from tests.transformers.test_modeling_common import (
@@ -36,15 +37,14 @@ from tests.transformers.test_modeling_common import (
     random_attention_mask,
 )
 
-# from parameterized import parameterized
-
 
 class Ernie4_5ModelTester:
     def __init__(
         self,
         parent,
         vocab_size=32000,
-        hidden_size=64,
+        hidden_size=1024,
+        head_dim=128,
         num_hidden_layers=2,
         num_attention_heads=8,
         masked_softmax_fusion=True,
@@ -76,6 +76,7 @@ class Ernie4_5ModelTester:
         self.parent: Ernie4_5ModelTest = parent
         self.vocab_size = vocab_size
         self.hidden_size = hidden_size
+        self.head_dim = head_dim
         self.num_hidden_layers = num_hidden_layers
         self.num_attention_heads = num_attention_heads
         self.masked_softmax_fusion = masked_softmax_fusion
@@ -434,115 +435,115 @@ class Ernie4_5GenerationD2STest(GenerationD2STestMixin, unittest.TestCase):
     internal_testing_model = "__internal_testing__/micro-random-ernie4_5"
 
 
-# class Ernie4_5CompatibilityTest(unittest.TestCase):
-#     test_model_id = "hf-internal-testing/tiny-random-Ernie4_5Model"
+class Ernie4_5CompatibilityTest(unittest.TestCase):
+    test_model_id = "hf-internal-testing/tiny-random-Ernie4_5Model"
 
-#     @classmethod
-#     @require_package("transformers", "torch")
-#     def setUpClass(cls) -> None:
-#         from transformers import Ernie4_5Config, Ernie4_5ForCausalLM
+    @classmethod
+    @require_package("transformers", "torch")
+    def setUpClass(cls) -> None:
+        from transformers import Ernie4_5Config, Ernie4_5ForCausalLM
 
-#         # when python application is done, `TemporaryDirectory` will be free
-#         cls.torch_model_path = tempfile.TemporaryDirectory().name
-#         config = Ernie4_5Config(hidden_size=16, num_hidden_layers=1, num_attention_heads=2)
-#         model = Ernie4_5ForCausalLM(config)
-#         model.save_pretrained(cls.torch_model_path)
+        # when python application is done, `TemporaryDirectory` will be free
+        cls.torch_model_path = tempfile.TemporaryDirectory().name
+        config = Ernie4_5Config(hidden_size=16, num_hidden_layers=1, num_attention_heads=2)
+        model = Ernie4_5ForCausalLM(config)
+        model.save_pretrained(cls.torch_model_path)
 
-#     @require_package("transformers", "torch")
-#     def test_ernie4_5_converter(self):
-#         # 1. create common input
-#         input_ids = np.random.randint(100, 200, [1, 20])
+    @require_package("transformers", "torch")
+    def test_ernie4_5_converter(self):
+        # 1. create common input
+        input_ids = np.random.randint(100, 200, [1, 20])
 
-#         # 2. forward the paddle model
-#         from paddleformers.transformers import Ernie4_5Model
+        # 2. forward the paddle model
+        from paddleformers.transformers import Ernie4_5Model
 
-#         paddle_model = Ernie4_5Model.from_pretrained(self.torch_model_path, convert_from_torch=True)
-#         paddle_model.eval()
-#         paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
+        paddle_model = Ernie4_5Model.from_pretrained(self.torch_model_path, convert_from_hf=True)
+        paddle_model.eval()
+        paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
 
-#         # 3. forward the torch  model
-#         import torch
-#         from transformers import Ernie4_5Model
+        # 3. forward the torch  model
+        import torch
+        from transformers import Ernie4_5Model
 
-#         torch_model = Ernie4_5Model.from_pretrained(self.torch_model_path)
-#         torch_model.eval()
-#         torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+        torch_model = Ernie4_5Model.from_pretrained(self.torch_model_path)
+        torch_model.eval()
+        torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-#         self.assertTrue(
-#             np.allclose(
-#                 paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                 torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                 rtol=1e-2,
-#             )
-#         )
+        self.assertTrue(
+            np.allclose(
+                paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                rtol=1e-2,
+            )
+        )
 
-#     @require_package("transformers", "torch")
-#     def test_ernie4_5_converter_from_local_dir(self):
-#         with tempfile.TemporaryDirectory() as tempdir:
+    @require_package("transformers", "torch")
+    def test_ernie4_5_converter_from_local_dir(self):
+        with tempfile.TemporaryDirectory() as tempdir:
 
-#             # 1. create common input
-#             input_ids = np.random.randint(100, 200, [1, 20])
+            # 1. create common input
+            input_ids = np.random.randint(100, 200, [1, 20])
 
-#             # 2. forward the torch  model
-#             import torch
-#             from transformers import Ernie4_5Model
+            # 2. forward the torch  model
+            import torch
+            from transformers import Ernie4_5Model
 
-#             torch_model = Ernie4_5Model.from_pretrained(self.torch_model_path)
-#             torch_model.eval()
-#             torch_model.save_pretrained(tempdir)
-#             torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+            torch_model = Ernie4_5Model.from_pretrained(self.torch_model_path)
+            torch_model.eval()
+            torch_model.save_pretrained(tempdir)
+            torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-#             # 2. forward the paddle model
-#             from paddleformers.transformers import Ernie4_5Model
+            # 2. forward the paddle model
+            from paddleformers.transformers import Ernie4_5Model
 
-#             paddle_model = Ernie4_5Model.from_pretrained(tempdir, convert_from_torch=True)
-#             paddle_model.eval()
-#             paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
+            paddle_model = Ernie4_5Model.from_pretrained(tempdir, convert_from_hf=True)
+            paddle_model.eval()
+            paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
 
-#             self.assertTrue(
-#                 np.allclose(
-#                     paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                     torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                     rtol=1e-2,
-#                 )
-#             )
+            self.assertTrue(
+                np.allclose(
+                    paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                    torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                    rtol=1e-2,
+                )
+            )
 
-#     @parameterized.expand([("Ernie4_5Model",), ("Ernie4_5ForCausalLM",)])
-#     @require_package("transformers", "torch")
-#     def test_ernie4_5_classes_from_local_dir(self, class_name, pytorch_class_name: str | None = None):
-#         pytorch_class_name = pytorch_class_name or class_name
-#         with tempfile.TemporaryDirectory() as tempdir:
+    @parameterized.expand([("Ernie4_5Model",), ("Ernie4_5ForCausalLM",)])
+    @require_package("transformers", "torch")
+    def test_ernie4_5_classes_from_local_dir(self, class_name, pytorch_class_name: str | None = None):
+        pytorch_class_name = pytorch_class_name or class_name
+        with tempfile.TemporaryDirectory() as tempdir:
 
-#             # 1. create common input
-#             input_ids = np.random.randint(100, 200, [1, 20])
+            # 1. create common input
+            input_ids = np.random.randint(100, 200, [1, 20])
 
-#             # 2. forward the torch model
-#             import torch
-#             import transformers
+            # 2. forward the torch model
+            import torch
+            import transformers
 
-#             torch_model_class = getattr(transformers, pytorch_class_name)
-#             torch_model = torch_model_class.from_pretrained(self.torch_model_path)
-#             torch_model.eval()
+            torch_model_class = getattr(transformers, pytorch_class_name)
+            torch_model = torch_model_class.from_pretrained(self.torch_model_path)
+            torch_model.eval()
 
-#             torch_model.save_pretrained(tempdir)
-#             torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+            torch_model.save_pretrained(tempdir)
+            torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-#             # 3. forward the paddle model
-#             from paddleformers import transformers
+            # 3. forward the paddle model
+            from paddleformers import transformers
 
-#             paddle_model_class = getattr(transformers, class_name)
-#             paddle_model = paddle_model_class.from_pretrained(tempdir, convert_from_torch=True)
-#             paddle_model.eval()
+            paddle_model_class = getattr(transformers, class_name)
+            paddle_model = paddle_model_class.from_pretrained(tempdir, convert_from_hf=True)
+            paddle_model.eval()
 
-#             paddle_logit = paddle_model(paddle.to_tensor(input_ids), return_dict=False)[0]
+            paddle_logit = paddle_model(paddle.to_tensor(input_ids), return_dict=False)[0]
 
-#             self.assertTrue(
-#                 np.allclose(
-#                     paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                     torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-#                     atol=1e-3,
-#                 )
-#             )
+            self.assertTrue(
+                np.allclose(
+                    paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                    torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+                    atol=1e-3,
+                )
+            )
 
 
 if __name__ == "__main__":
