+++ tests/transformers/llama/test_modeling.py
@@ -14,15 +14,16 @@
 # limitations under the License.
 from __future__ import annotations
 
-import tempfile
+# import tempfile
 import unittest
 
-import numpy as np
+# import numpy as np
 import paddle
-from parameterized import parameterized
 
 from paddleformers.transformers import LlamaConfig, LlamaForCausalLM, LlamaModel
-from tests.testing_utils import require_package, slow
+
+# from tests.testing_utils import require_package, slow
+from tests.testing_utils import slow
 from tests.transformers.test_configuration_common import ConfigTester
 from tests.transformers.test_generation_utils import GenerationTesterMixin
 from tests.transformers.test_modeling_common import (
@@ -32,6 +33,8 @@ from tests.transformers.test_modeling_common import (
     ids_tensor,
     random_attention_mask,
 )
+
+# from parameterized import parameterized
 
 
 class LlamaModelTester:
@@ -394,115 +397,115 @@ class LlamaGenerationD2STest(GenerationD2STestMixin, unittest.TestCase):
     internal_testing_model = "__internal_testing__/micro-random-llama"
 
 
-class LlamaCompatibilityTest(unittest.TestCase):
-    test_model_id = "hf-internal-testing/tiny-random-LlamaModel"
+# class LlamaCompatibilityTest(unittest.TestCase):
+#     test_model_id = "hf-internal-testing/tiny-random-LlamaModel"
 
-    @classmethod
-    @require_package("transformers", "torch")
-    def setUpClass(cls) -> None:
-        from transformers import LlamaConfig, LlamaForCausalLM
+#     @classmethod
+#     @require_package("transformers", "torch")
+#     def setUpClass(cls) -> None:
+#         from transformers import LlamaConfig, LlamaForCausalLM
 
-        # when python application is done, `TemporaryDirectory` will be free
-        cls.torch_model_path = tempfile.TemporaryDirectory().name
-        config = LlamaConfig(hidden_size=16, num_hidden_layers=1, num_attention_heads=2)
-        model = LlamaForCausalLM(config)
-        model.save_pretrained(cls.torch_model_path)
+#         # when python application is done, `TemporaryDirectory` will be free
+#         cls.torch_model_path = tempfile.TemporaryDirectory().name
+#         config = LlamaConfig(hidden_size=16, num_hidden_layers=1, num_attention_heads=2)
+#         model = LlamaForCausalLM(config)
+#         model.save_pretrained(cls.torch_model_path)
 
-    @require_package("transformers", "torch")
-    def test_llama_converter(self):
-        # 1. create common input
-        input_ids = np.random.randint(100, 200, [1, 20])
+#     @require_package("transformers", "torch")
+#     def test_llama_converter(self):
+#         # 1. create common input
+#         input_ids = np.random.randint(100, 200, [1, 20])
 
-        # 2. forward the paddle model
-        from paddleformers.transformers import LlamaModel
+#         # 2. forward the paddle model
+#         from paddleformers.transformers import LlamaModel
 
-        paddle_model = LlamaModel.from_pretrained(self.torch_model_path, convert_from_torch=True)
-        paddle_model.eval()
-        paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
+#         paddle_model = LlamaModel.from_pretrained(self.torch_model_path, convert_from_torch=True)
+#         paddle_model.eval()
+#         paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
 
-        # 3. forward the torch  model
-        import torch
-        from transformers import LlamaModel
+#         # 3. forward the torch  model
+#         import torch
+#         from transformers import LlamaModel
 
-        torch_model = LlamaModel.from_pretrained(self.torch_model_path)
-        torch_model.eval()
-        torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+#         torch_model = LlamaModel.from_pretrained(self.torch_model_path)
+#         torch_model.eval()
+#         torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-        self.assertTrue(
-            np.allclose(
-                paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                rtol=1e-2,
-            )
-        )
+#         self.assertTrue(
+#             np.allclose(
+#                 paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                 torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                 rtol=1e-2,
+#             )
+#         )
 
-    @require_package("transformers", "torch")
-    def test_llama_converter_from_local_dir(self):
-        with tempfile.TemporaryDirectory() as tempdir:
+#     @require_package("transformers", "torch")
+#     def test_llama_converter_from_local_dir(self):
+#         with tempfile.TemporaryDirectory() as tempdir:
 
-            # 1. create common input
-            input_ids = np.random.randint(100, 200, [1, 20])
+#             # 1. create common input
+#             input_ids = np.random.randint(100, 200, [1, 20])
 
-            # 2. forward the torch  model
-            import torch
-            from transformers import LlamaModel
+#             # 2. forward the torch  model
+#             import torch
+#             from transformers import LlamaModel
 
-            torch_model = LlamaModel.from_pretrained(self.torch_model_path)
-            torch_model.eval()
-            torch_model.save_pretrained(tempdir)
-            torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+#             torch_model = LlamaModel.from_pretrained(self.torch_model_path)
+#             torch_model.eval()
+#             torch_model.save_pretrained(tempdir)
+#             torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-            # 2. forward the paddle model
-            from paddleformers.transformers import LlamaModel
+#             # 2. forward the paddle model
+#             from paddleformers.transformers import LlamaModel
 
-            paddle_model = LlamaModel.from_pretrained(tempdir, convert_from_torch=True)
-            paddle_model.eval()
-            paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
+#             paddle_model = LlamaModel.from_pretrained(tempdir, convert_from_torch=True)
+#             paddle_model.eval()
+#             paddle_logit = paddle_model(paddle.to_tensor(input_ids))[0]
 
-            self.assertTrue(
-                np.allclose(
-                    paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                    torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                    rtol=1e-2,
-                )
-            )
+#             self.assertTrue(
+#                 np.allclose(
+#                     paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                     torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                     rtol=1e-2,
+#                 )
+#             )
 
-    @parameterized.expand([("LlamaModel",), ("LlamaForCausalLM",)])
-    @require_package("transformers", "torch")
-    def test_llama_classes_from_local_dir(self, class_name, pytorch_class_name: str | None = None):
-        pytorch_class_name = pytorch_class_name or class_name
-        with tempfile.TemporaryDirectory() as tempdir:
+#     @parameterized.expand([("LlamaModel",), ("LlamaForCausalLM",)])
+#     @require_package("transformers", "torch")
+#     def test_llama_classes_from_local_dir(self, class_name, pytorch_class_name: str | None = None):
+#         pytorch_class_name = pytorch_class_name or class_name
+#         with tempfile.TemporaryDirectory() as tempdir:
 
-            # 1. create common input
-            input_ids = np.random.randint(100, 200, [1, 20])
+#             # 1. create common input
+#             input_ids = np.random.randint(100, 200, [1, 20])
 
-            # 2. forward the torch model
-            import torch
-            import transformers
+#             # 2. forward the torch model
+#             import torch
+#             import transformers
 
-            torch_model_class = getattr(transformers, pytorch_class_name)
-            torch_model = torch_model_class.from_pretrained(self.torch_model_path)
-            torch_model.eval()
+#             torch_model_class = getattr(transformers, pytorch_class_name)
+#             torch_model = torch_model_class.from_pretrained(self.torch_model_path)
+#             torch_model.eval()
 
-            torch_model.save_pretrained(tempdir)
-            torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
+#             torch_model.save_pretrained(tempdir)
+#             torch_logit = torch_model(torch.tensor(input_ids), return_dict=False)[0]
 
-            # 3. forward the paddle model
-            from paddleformers import transformers
+#             # 3. forward the paddle model
+#             from paddleformers import transformers
 
-            paddle_model_class = getattr(transformers, class_name)
-            paddle_model = paddle_model_class.from_pretrained(tempdir, convert_from_torch=True)
-            paddle_model.eval()
+#             paddle_model_class = getattr(transformers, class_name)
+#             paddle_model = paddle_model_class.from_pretrained(tempdir, convert_from_torch=True)
+#             paddle_model.eval()
 
-            paddle_logit = paddle_model(paddle.to_tensor(input_ids), return_dict=False)[0]
+#             paddle_logit = paddle_model(paddle.to_tensor(input_ids), return_dict=False)[0]
 
-            self.assertTrue(
-                np.allclose(
-                    paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                    torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
-                    atol=1e-3,
-                )
-            )
+#             self.assertTrue(
+#                 np.allclose(
+#                     paddle_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                     torch_logit.detach().cpu().reshape([-1])[:9].numpy(),
+#                     atol=1e-3,
+#                 )
+#             )
 
 
 if __name__ == "__main__":
